+++ src/components/TimeRange/TimeRange.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { instanceOf, string } from 'prop-types';
 import classNames from 'classnames';
 import { isSameDay, formatDateIntoPartials } from '../../util/dates';
-import { injectIntl, intlShape } from '../../util/reactIntl';
+import { useIntl } from '../../util/reactIntl';
 import { DATE_TYPE_DATE, DATE_TYPE_TIME, DATE_TYPE_DATETIME, propTypes } from '../../util/types';
 
 import css from './TimeRange.module.css';
@@ -13,7 +13,8 @@ const DASH = '–';
 const BREAK_WORD_MIN_LENGTH = 27;
 
 export const TimeRangeComponent = props => {
-  const { rootClassName, className, startDate, endDate, dateType, timeZone, intl } = props;
+  const intl = useIntl();
+  const { rootClassName, className, startDate, endDate, dateType, timeZone } = props;
   const start = formatDateIntoPartials(startDate, intl, { timeZone });
   const end = formatDateIntoPartials(endDate, intl, { timeZone });
   const isSingleDay = isSameDay(startDate, endDate, timeZone);
@@ -100,9 +101,6 @@ TimeRangeComponent.propTypes = {
   endDate: instanceOf(Date).isRequired,
   dateType: propTypes.dateType,
   timeZone: string,
-
-  // from injectIntl
-  intl: intlShape.isRequired,
 };
 
-export default injectIntl(TimeRangeComponent);
+export default TimeRangeComponent;